requirements.txt
pyyaml
pyspark
requests
requests_oauthlib
pandas
pika
jupyterlab
